Remove `dataset` module
Moved the single function `load` from `dataset` to `__main__`.
